modified to remove extraneous SSLModes from MaxwellMysqlConfig
